requirements.txt
boto3==1.21.19
botocore==1.24.19
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Flask==2.2.2
GPUtil==1.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
numpy==1.23.3
Pillow==9.2.0
psutil==5.9.2
py-cpuinfo==8.0.0
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.8
Werkzeug==2.2.2
